+++ mutations/page_title_body_class.pl
@@ -1,0 +1,1 @@
+s/return \$this->hidetitle \? \[self::NOTITLE_CLASS\] : \[\];/return [];/;