+++ mutations/page_title_heading.pl
@@ -1,0 +1,1 @@
+s/return \$this->hidetitle \? '' : get_string\('pluginname', 'block_compass'\);/return get_string('pluginname', 'block_compass');/;