+++ mutations/page_title_hidden_h1.pl
@@ -1,0 +1,1 @@
+s/<h1 class="visually-hidden">/<h1 class="d-none">/;