+++ mutations/page_title_setting.pl
@@ -1,0 +1,1 @@
+s/get_config\('block_compass', 'hide_page_title'\) === 1;/get_config('block_compass', 'hide_page_title') !== 1;/;